PYTHON_generate_label_QRCode/requirements.txt
pillow
qrcode
requests
fastapi[standard]
opencv-contrib-python
matplotlib